pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuega-ai"
version = "1.0.0"
description = "AI-powered full-service digital agency for LATAM small businesses"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "anthropic>=0.39.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "tweepy>=4.14.0",
    "resend>=2.0.0",
    "redis>=5.0.0",
    "websockets>=12.0",
    "rich>=13.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5.0"]

[tool.setuptools.packages.find]
include = ["backend*"]
